+++ py-polars/tests/unit/lazyframe/test_rename.py
@@ -6,3 +6,10 @@ def test_lazy_rename() -> None:
 
     result = df.lazy().rename({"y": "x", "x": "y"}).select(["x", "y"])
     assert result.collect().to_dict(as_series=False) == {"x": [2], "y": [1]}
+
+
+def test_remove_redundant_mapping_4668() -> None:
+    lf = pl.LazyFrame([["a"]] * 2, ["A", "B "]).lazy()
+    clean_name_dict = {x: " ".join(x.split()) for x in lf.columns}
+    lf = lf.rename(clean_name_dict)
+    assert lf.columns == ["A", "B"]
