+++ py-polars/tests/unit/operations/aggregation/test_vertical.py
@@ -74,3 +74,11 @@ def test_mean_overflow() -> None:
 
     result = df.with_columns(pl.col("value").cast(pl.Int32)).get_column("value").mean()
     assert np.isclose(result, expected)
+
+
+def test_deep_subexpression_f32_schema_7129() -> None:
+    df = pl.DataFrame({"a": [1.1, 2.3, 3.4, 4.5]}, schema={"a": pl.Float32()})
+    assert df.with_columns(pl.col("a") - pl.col("a").median()).dtypes == [pl.Float32]
+    assert df.with_columns(
+        (pl.col("a") - pl.col("a").mean()) / (pl.col("a").std() + 0.001)
+    ).dtypes == [pl.Float32]
